Reuse S3 code and delete image after test
