+++ test/assets/sample-image.jpg
